Add support for LLVM Debug Info metadata
This extends the AST to support the DWARF-based debug info metadata that
are produced by C programs run through clang-3.8. No guarantees are made
for other source languages, as this is not a comprehensive
implementation of all the features supported by LLVM's parser.

The big TODO with what is here is to turn some of the DWARF constants
into proper sum types so we can handle and pretty-print them using their
string names.