Fix rendering call function signatures in certain cases.
When calling a function in which the function value is not a symbol (for
example, is a `ValConstExpr` due to `bitcast`ing) the rendering was not
correct.

The rendering was also incorrect in cases of var args and functions that
return function pointers.  In these cases the function arguments must
also appear as a parameter to call.  Rather than be selective to just
these case this patch opts to always include the args.