Support `fneg` instruction
Recent versions of LLVM represent floating-point negation with a dedicated
`fneg` instruction, as discovered in GaloisInc/llvm-pretty-bc-parser#148.
`fneg` is considered to be a unary operation, much like the `add`, `sub`, etc.
instructions are binary operations. As a result, rather than adding a dedicated
`fneg` instruction, I added a `UnaryArith` instruction, which acts like
(binary) `Arith` but with a single parameter. The `UnaryArithOp` data type,
then, only consists of `FNeg` for now, but there could be additional operations
in future versions of LLVM.