Add PtrOpaque type
This adds basic support for representing and pretty-printing opaque `ptr`
types, which were introduced in LLVM 13 as part of a longer-term plan to stop
using non-opaque pointer types (e.g., `i8*`). See
https://llvm.org/docs/OpaquePointers.html for further details.

The next step in `llvm-pretty` is to make sure that all of the instructions
that involve pointers in some way (e.g., `call`) have all the information that
they need to work with opaque pointers. I will be addressing this in subsequent
commits.

This is one step towards #102.