Handle function types correctly when pretty-printing call(br)
Previously, the code in `ppCallSym` assumed that the function type stored in a
`call` instruction was a pointer type, but this should actually be a raw
function type. See the discussion at
https://github.com/GaloisInc/llvm-pretty-bc-parser/issues/189#issuecomment-1476068137
for more on this.  We now assume the convention of raw function types in `call`
instructions, as well as the related `callbr` and `invoke` instructions.

This addresses one part of #102.