Store load type separately in Load
With opaque pointers, one cannot tell what the type being loaded is by
inspecting the pointer in the `load` instruction. As a result, we now store the
load type separately in the `Load` instruction so that it can be determined
regardless of whether opaque pointers are used or not.

See #102.