Store basis type separately in GEP/ConstGEP
With opaque pointers, one cannot tell what the basis type for a `getelementptr`
instruction (or constant expression) is by inspecting the parent pointer. As a
result, we now store the basis type separately in `GEP`/`ConstGEP` so that it
can be determined regardless of whether opaque pointers are used or not.

This also requires tweaking the types of the `resolveGepFull` and `resolveGep`
functions to make the basis type separate from the parent pointer type.

Because this requires making a backwards-incompatible change to `ConstGEP`, I
took the opportunity to include the parent pointer value as a distinguished
field in `ConstGEP`. The `GEP` data constructor already does this, and it seems
oddly asymmetric to not have `ConstGEP` do the same, especially since LLVM
requires it to be present.

See #102.