Make dicuRangesBaseAddress a Maybe Bool
As the LLVM source code indicates in
https://github.com/llvm/llvm-project/blob/9841daf27076886c6ab8e155eb812bda76f77532/llvm/lib/AsmParser/LLParser.cpp#L5215,
the intent was always to have `dicuRangesBaseAddress` be an optional field.
This updates the `llvm-pretty` AST to reflect this. With this change, picking a
default value of `Nothing` will cause `dicuRangesBaseAddress` not to be
pretty-printed, which makes it easier to support pre-11 versions of LLVM (when
`rangesBaseAddress` was introduced).

Fixes #113.