Support non-constant sizes/offsets in DI{Basic,Derived,Composite}Type
This is a feature that was introduced in LLVM 21. For the most part, this is
accomplished by changed the types of certain type-related debug nodes from
`Word64` to `Maybe (ValMd' lab)`. This also required adding a `lab` type
parameter to `DIBasicType`, which has some knock-on code changes elsewhere to
deal with the extra type parameter.

Towards https://github.com/GaloisInc/llvm-pretty-bc-parser/issues/321.